Redone with no-demon thread as current pytest fixtures do not like these Extra test folder deleted
 On branch fix_moved_records_not_cleaned
 Changes to be committed:
	modified:   src/watchdog/observers/inotify_c.py
	modified:   tests/test_inotify_c.py
	deleted:    tests/threaded/conftest.py
	deleted:    tests/threaded/test_inotify_c.py
